# gh-117609: Speedup `os.path`