--- a/Lib/ntpath.py
+++ b/Lib/ntpath.py
@@ -784,7 +784,6 @@ supports_unicode_filenames = True
 
 def relpath(path, start=None):
     """Return a relative version of a path"""
-
     path = os.fspath(path)
     if not path:
         raise ValueError("no path specified")