--- a/Lib/posixpath.py
+++ b/Lib/posixpath.py
@@ -348,7 +348,7 @@ except ImportError:
             sep = '/'
             dot = '.'
             dotdot = '..'
-        if not path:
+        if not path or path == dot:
             return dot
         _, initial_slashes, path = splitroot(path)
         comps = path.split(sep)
@@ -500,10 +500,12 @@ def relpath(path, start=None):
         curdir = b'.'
         sep = b'/'
         pardir = b'..'
+        getcwd = os.getcwdb
     else:
         curdir = '.'
         sep = '/'
         pardir = '..'
+        getcwd = os.getcwd
 
     if start is None:
         start = curdir
@@ -511,10 +513,12 @@ def relpath(path, start=None):
         start = os.fspath(start)
 
     try:
-        start_rest = abspath(start).lstrip(sep)
-        path_rest = abspath(path).lstrip(sep)
-        start_list = start_rest.split(sep) if start_rest else []
-        path_list = path_rest.split(sep) if path_rest else []
+        start_abs = getcwd() if not start or start == curdir else abspath(start)
+        path_abs = abspath(path)
+        start_tail = start_abs.lstrip(sep)
+        path_tail = path_abs.lstrip(sep)
+        start_list = start_tail.split(sep) if start_tail else []
+        path_list = path_tail.split(sep) if path_tail else []
         # Work out how much of the filepath is shared by start and path.
         i = len(commonprefix([start_list, path_list]))