--- a/Misc/NEWS.d/next/Core and Builtins/2024-04-07-19-03-53.gh-issue-117609.J-C6J2.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2024-04-07-19-03-53.gh-issue-117609.J-C6J2.rst
@@ -1,0 +1,1 @@
+Speedup some functions in :mod:`os.path`.