Benchmark:

- `ntpath.py`
    ```bat
    ::test.bat
    @echo off
    echo join() && python -m timeit -s "import before.ntpath" "before.ntpath.join('foo')" && python -m timeit -s "import after.ntpath" "after.ntpath.join('foo')"
    echo split() && python -m timeit -s "import before.ntpath" "before.ntpath.split('foo')" && python -m timeit -s "import after.ntpath" "after.ntpath.split('foo')"
    echo ismount() && python -m timeit -s "import before.ntpath" "before.ntpath.ismount('//')" && python -m timeit -s "import after.ntpath" "after.ntpath.ismount('//')"
    echo expanduser() && python -m timeit -s "import before.ntpath" "before.ntpath.expanduser('~')" && python -m timeit -s "import after.ntpath" "after.ntpath.expanduser('~')"
    echo expanduser('~' + 'a' * 100) && python -m timeit -s "import before.ntpath" "before.ntpath.expanduser('~' + 'a' * 100)" && python -m timeit -s "import after.ntpath" "after.ntpath.expanduser('~' + 'a' * 100)"
    ::echo realpath() && python -m timeit -s "import before.ntpath" "before.ntpath.realpath('foo')" && python -m timeit -s "import after.ntpath" "after.ntpath.realpath('foo')"
    ::echo realpath('C:/') && python -m timeit -s "import before.ntpath" "before.ntpath.realpath('C:/')" && python -m timeit -s "import after.ntpath" "after.ntpath.realpath('C:/')"
    ```

    ```none
    join()
    500000 loops, best of 5: 693 nsec per loop # before
    500000 loops, best of 5: 650 nsec per loop # after
    # -> 1.07x faster
    split()
    500000 loops, best of 5: 866 nsec per loop # before
    500000 loops, best of 5: 799 nsec per loop # after
    # -> 1.08x faster
    ismount()
    200000 loops, best of 5: 1.19 usec per loop # before
    200000 loops, best of 5: 1.13 usec per loop # after
    # -> 1.05x faster
    expanduser()
    500000 loops, best of 5: 973 nsec per loop # before
    500000 loops, best of 5: 967 nsec per loop # after
    # -> no difference
    expanduser('~' + 'a' * 100)
    20000 loops, best of 5: 18 usec per loop # before
    20000 loops, best of 5: 11.2 usec per loop # after
    # -> 1.61x faster (for long users)
    ```
- `posixpath.py`
    **TODO**

<!-- gh-issue-number: gh-117609 -->
* Issue: gh-117609
<!-- /gh-issue-number -->