@erlend-aasland 2024-04-08
Note that in general, we do not accept PRs with minor speedups like this. Moreover, any code change risks the introduction of new bugs, as the currently failing CI indicates.