@nineteendo 2024-04-08
Also note that this is still a work in progress:

1. I have not yet done benchmarks on Unix, so I don't know how large the speedups are there.
2. I'll be investigating the failing tests. They're most likely related to running `abspath()` of the other platform as the other changes don't rely on any assumptions.