@nineteendo 2024-04-08
Also, in the old code this loop calls `_get_bothseps()` in **every iteration** without caching the result:

https://github.com/python/cpython/blob/e338e1a4ec5e43a02447f4ec80320d7fc12b3ed4/Lib/ntpath.py#L377-L378

For long users (>100 characters) the new code is 1.61x faster, which is not a minor speedup.